% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{Swedish county data with short name}
\format{
\subsection{\code{counties}}{

A named character vector of length 21
\describe{
\item{name(counties)}{Full name of county}
\item{counties}{The one or two letter county shortname}

}
}
}
\source{
\url{https://sv.wikipedia.org/wiki/L\%C3\%A4nsbokstav}
}
\usage{
data(counties)
}
\description{
Swedish county data with short name
}
\keyword{datasets}
