% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGenoType.R
\name{CreateGenotype}
\alias{CreateGenotype}
\title{Extract most likely genotype data from replicates}
\usage{
CreateGenotype(
  file = file,
  locus = c("Locus1", "Locus2"),
  naStrings = c("NA", "-99", "0", "000", "No peaks in locus", "No peaks"),
  notAllowed = c("Unbinned peaks", "Too many alleles", "Unbinned peaks in locus",
    "Peaks outside loci")
)
}
\arguments{
\item{file}{name of the file with replicate data}

\item{locus}{names of the loci to be analysed}

\item{naStrings}{string treated as NA in import}

\item{notAllowed}{strings that will stop import}
}
\value{
a list of three dataframes: Genotypes, QC and number of
missing data points
}
\description{
Function to extract the most likely genotype scores from multiple
replicates of genotype data. The structure of the input data is a
table (excel etc) where the first column contains sample names and
the genotypes of interest can be in any other column in the input
data. Note that the function assumes that replicates will have the
same name after removing the tail part of the original name. Eg. if
replicates are named SEP1000001_mix1_rep1, SEP1000001_mix1_rep2 etc
the sample name will be SEP1000001.
NB! The naming used for loci in this implementation assumes that
there are spaces in the names. Hence the need to modify the names
with the replacement of "..." with ".".
}
