% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateImport}
\alias{CreateImport}
\title{Function to from the match database create import file according
the standard given by rovbase}
\usage{
CreateImport(fromMatchDB, RBData)
}
\arguments{
\item{fromMatchDB}{Dataframe with target data}

\item{RBData}{Dataframe with metadata from rovbase for the samples
in fromMatchDB}
}
\value{
a dataframe with the neccessary content for import to
rovbase.
}
\description{
Function to from the match database create import file according
the standard given by rovbase
}
