% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInputData.R
\name{CreateSampleImport}
\alias{CreateSampleImport}
\title{Function to from the match database and rovbasefiles create an import file according
the standard given by rovbase}
\usage{
CreateSampleImport(
  matchDBFile = "matchdb.csv",
  metaDataFile = "rovbasemetadata.xlsx",
  year = 24,
  onlyNew = TRUE,
  sampleType = "Spillning",
  methodID = 18,
  finalVersion = FALSE
)
}
\arguments{
\item{matchDBFile}{file (.xlsx, .csv, tsv) with data to be imported. This are
samples matched to local genetic database and that are to be imported to
rovbase. The file should have the following columns (but can contain others):
\itemize{
\item index
\item date
\item north
\item east
\item gender
\item date_changed
\item confirmed_dead
\item individ
}}

\item{metaDataFile}{file with metadata from rovbase. This data will be merged
with samples informtation from the matchDBFile found in the fromMatchDB dataframe.
\strong{NB! samples found in the dataframe matchDBFile that is not found in rovbase
will be saved in output of this function.}
This dataframe needs to contain the following columns (but can contain others):
\itemize{
\item Strekkodeprove - SEP or M number. Original name in rb = "Strekkode (Prøve)"
\item DNAIDProve - DNA ID. Original name in rb = "DNAID (Prøve)"
\item RovbaseIDProver - Rovbase ID. Original name in rb = "RovbaseID (Prøve)"
\item Individ - Rovbase individ from rb.
\item KjonnIndivid - Rovbase sex. Original name in rb = "Kjønn (Individ)"
\item Provetype - Sample type. Original name in rb = "Prøvetype"
}}

\item{year}{which years were the samples collected in NB! accepts the last two
digits so 2024 should be 24.}

\item{onlyNew}{Since new individuals also need to have a trivial id created this
new inds should be run with this set to TRUE and samples that matched already
known individuals will be handled correctly if this is set to FALSE.}

\item{sampleType}{string setting the samples types that will be kept in the
output. NB! We mostly do scat samples and this should then be left as "Spillning".}

\item{methodID}{Number corresponding to the methods used for analyzing the samples.
8 SSR markers that were originally used at NRM for sample identification goes under
the methodID 18.}

\item{finalVersion}{Boolean value if set to TRUE only the columns needed for import
file to rovbase are saved. If FALSE a version with columns that can be manually checked
are retained. NB! Use FALSE and manually clean the file in excel unless you are 100\% sure
that the data is consistent in the matching data and rovbase.}
}
\value{
a dataframe with the neccessary content for import to
rovbase.

List with two dataframes. The one named miss will contain the samples
found in the matching database, but than are not found in rovbase. The second one
contains as dataframe with data almost ready for import to rovbase.
NB! The sex returned as "rbKjonnID" is the sex of the \strong{individual} in rovbase
and hence not actual extracted from the specific sample eg. matching with
"StrekkodeProve".
}
\description{
Function to from the match database and rovbasefiles create an import file according
the standard given by rovbase
}
