% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExtractNames}
\alias{ExtractNames}
\title{Function that extract individual names from index (SEP) or M numbers}
\usage{
ExtractNames(nameVector)
}
\arguments{
\item{nameVector}{a vector containing entries with names that have
replicate name and SSR mix attached. The latter will be stripped in
and only the \emph{actual} names of samples will be retained.}
}
\value{
a vector of names that 10 long if SEP or 7 long in all
other cases
}
\description{
Function that extract individual names from index (SEP) or M numbers
}
