% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExtractNames2}
\alias{ExtractNames2}
\title{Function that extract individual names from index (SEP) or M numbers}
\usage{
ExtractNames2(namesVector, sep = "_")
}
\arguments{
\item{nameVector}{a vector containing entries with names that have
replicate name and SSR mix attached, but separated by a character. Only
the string corresponding to the first part of the entry will be
retained eg. samplename_mix1_rep1 will return samplename.}
}
\value{
a vector of names
}
\description{
Function that extract individual names from index (SEP) or M numbers
}
