% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateTrivialID}
\alias{generateTrivialID}
\title{Generate trivial named based on the year that the individual was found and the origin.
NB! Only scat samples should be named with this and dead bears should instead have the SVA-ID
as trivial name.}
\usage{
generateTrivialID(
  importData,
  individualColumn = "individ",
  countyColumn = "Fylke",
  year
)
}
\arguments{
\item{importData}{dataframe with samples to be imported}

\item{individualColumn}{name of the column with individualdata}

\item{countyColumn}{name of the column with county data. NB! this should be as the names are given in
rovbase.}

\item{year}{what year are samples from.}
}
\value{
dataframe with two columns. One with individual name and the second containing the trivialname.
the trivial names will look like this from a sample from Norrbotten in 2024: BD24-001.
}
\description{
Generate trivial named based on the year that the individual was found and the origin.
NB! Only scat samples should be named with this and dead bears should instead have the SVA-ID
as trivial name.
}
