% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GenerateValidNames}
\alias{GenerateValidNames}
\title{Generate valid dataframe with "proper" names}
\usage{
GenerateValidNames(x)
}
\arguments{
\item{x}{a dataframe with names containing the names}
}
\value{
a vector of names that are valid and contain no strange characters.
}
\description{
Will generate a dataframe where column names have been cleaned up.
It will remove any white spaces and other illegal characters for names.
Will also remove any language specific characters that is not found in the
"latin-ascii" character set.
}
