% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LoadData}
\alias{LoadData}
\title{Helper function read facilitate data import of different file
formats}
\usage{
LoadData(
  filePath,
  sheet = 1,
  naStrings = c("NA", "-99", "0", "000", "No peaks in locus", "No peaks")
)
}
\arguments{
\item{sheet}{which sheet to read data from in excel and ods file formats}

\item{naStrings}{entries should be converted to NA values on import
\strong{NB! The function uses file endings to guess file formats. Will not work if file ending}
\strong{is not consistent with actual file type}}

\item{file_path}{name of file to import data from (with path)}
}
\value{
a dataframe of the imported data
}
\description{
Helper function read facilitate data import of different file
formats
}
