% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergePlate}
\alias{MergePlate}
\title{Add genotype results from SSR markers that are analysed on
different PCR mixes to a single dataframe}
\usage{
MergePlate(data1, data2)
}
\arguments{
\item{data1}{First dataframe with results}

\item{data2}{Second dataframe with results}
}
\description{
Add genotype results from SSR markers that are analysed on
different PCR mixes to a single dataframe
}
