% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergePlateS}
\alias{MergePlateS}
\title{Add genotype results stored in different dataframes into a single
dataframe. Note that the row names of the genotypes need to match.}
\usage{
MergePlateS(...)
}
\arguments{
\item{...}{dataframes to be merged.}
}
\value{
data frame with all genotypes results
}
\description{
Add genotype results stored in different dataframes into a single
dataframe. Note that the row names of the genotypes need to match.
}
