% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergePlateSum}
\alias{MergePlateSum}
\title{Add genotype results different marker types to
a single dataframe}
\usage{
MergePlateSum(data1, data2)
}
\arguments{
\item{data1}{First dataframe with results}

\item{data2}{Second dataframe with results}
}
\description{
Add genotype results different marker types to
a single dataframe
}
