% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MinMod}
\alias{MinMod}
\title{Extract the min value from a range of values. Returns NA without a
warning if all data is NA.}
\usage{
MinMod(x)
}
\arguments{
\item{x}{vector with values}
}
\value{
the minimum value of a vector or NA if only missing data is found
}
\description{
Extract the min value from a range of values. Returns NA without a
warning if all data is NA.
}
