% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{QCRep}
\alias{QCRep}
\title{Check if allele sizes from replicate runs are the same. Returns 'OK'
if all looks identical and returns 'Check' if replicates are
inconsistent. Returns "No data" if there is no data available.}
\usage{
QCRep(data = data, locus = "")
}
\arguments{
\item{data}{with genotyping results}

\item{locus}{of the marker to analysed}
}
\description{
Check if allele sizes from replicate runs are the same. Returns 'OK'
if all looks identical and returns 'Check' if replicates are
inconsistent. Returns "No data" if there is no data available.
}
