% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGenoType.R
\name{sexdetermination}
\alias{sexdetermination}
\title{Determine sex from genotype data}
\usage{
sexdetermination(data = data, locus = c("Male", "UaY15020", "UarY369.4"))
}
\arguments{
\item{data}{a dataframe with genotype results}
}
\value{
list with input data plus QC score and a vector with most likely sex
}
\description{
This function will based on the selected number of loci determine the
most likely sex of a sample from genotype data.
NB! The function assumes that the marker with name Male have the following
properties:
\enumerate{
\item 152 152 == Female
\item 98 152 == Male
\item 98  98 == Male
\item NA 152 == Female
\item NA  98 == Male
\item NA  NA == NA
}
}
\details{
The other markers that are y-chromosome markers have the following properties:
\enumerate{
\item NA NA == Female
\item Any other value == Male
}

Consensus for this is the extracted as follows:
\enumerate{
\item All marker same sex this is the true sex of the sample
\item All markers have NA, the sex is not determined NA is reported
\item In case of mismatch the most common sex i extracted.
\item The Male marker is better as the y-chromosome markers can not separate failed from female.
}
}
