% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WriteImportFile}
\alias{WriteImportFile}
\title{Creates/Writes excel workbook with pre-specified header read from file
combined with data from selected dataframe.}
\usage{
WriteImportFile(
  headerInfoFile,
  data,
  sheet = 1,
  startRow = 24,
  outputFile = "toimportfinal.xlsx"
)
}
\arguments{
\item{headerInfoFile}{excel file with headers to be part of output}

\item{data}{dataframe with data to be appended to the headerInfoFile}

\item{sheet}{which sheet in the excel file to write to}

\item{startRow}{integer corresponding to row in excel that dataframe will be written to}

\item{outputFile}{file name to write data to. Should have the file ending .xlsx}
}
\value{
Nothing, but will write and overwrite excel file on disk as a side effect.
}
\description{
Creates/Writes excel workbook with pre-specified header read from file
combined with data from selected dataframe.
}
