% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExtractNames}
\alias{ExtractNames}
\title{Extract basename of string}
\usage{
ExtractNames(x, sep = "_")
}
\arguments{
\item{x}{a vector containing entries with names that where
the names contain the character @param sep that will be used to
split the names. Only characters or number prior to the first
instance of @param sep will be retained.}

\item{sep}{character to split input name on.}
}
\value{
a vector of names that correspond to leading part of names.
}
\description{
Seperates vector of character strings on @param sep
and returns a vector where every entry contains all characters
occuring prior to @sep in the input vector.
}
