% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MaxMod}
\alias{MaxMod}
\title{Extract the max value from a range of values.}
\usage{
MaxMod(x)
}
\arguments{
\item{x}{vector with values}
}
\value{
the maximum value from a vector or NA if only missing data is found
}
\description{
Returns maximum value from a vector of values and
returns NA if all data is NA.
}
