% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergePlateS}
\alias{MergePlateS}
\title{Merges data from multiple dataframes to single dataframe}
\usage{
MergePlateS(...)
}
\arguments{
\item{...}{list of data ot merged}
}
\value{
datafram with merged data
}
\description{
accepts any number of dataframes as arguments and if
row names are the same they will be merged a single dataframe.
}
