% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{QCRep}
\alias{QCRep}
\title{Quality control of  allele sizes from replicate runs.}
\usage{
QCRep(data = data, locus = "")
}
\arguments{
\item{data}{with genotyping results}

\item{locus}{of the marker to be analysed}
}
\description{
Quality controls will compare all replicate run of
sample and return 'OK' if all values are identical excluding
missing data.  If values are inconsistent it will return
'Check'. If there is only missing data on a given sample it
will return 'No data'
}
